Update trainSim.md
Made it clear the other game mods referenced didn't use physics in their trains.

Arguably Create has an additive mod to add physics to their trains, but it's a popular subject in the Create community to also say that very mod is a source of many game-breaking bugs, so I do not count it.